Modularization: Moving Value.toOption/.toTry to Option and Try. Removing Future.toTry in favor of await().getValue().get().
