+++ vavr/src/main/java/io/vavr/control/Option.java
@@ -20,7 +20,6 @@
 package io.vavr.control;
 
 import io.vavr.PartialFunction;
-import io.vavr.Tuple;
 import io.vavr.Value;
 import io.vavr.collection.Iterator;
 import io.vavr.collection.Seq;
@@ -430,6 +429,27 @@ public interface Option<T> extends Value<T>, Serializable {
     default <U> U transform(Function<? super Option<T>, ? extends U> f) {
         Objects.requireNonNull(f, "f is null");
         return f.apply(this);
+    }
+
+    /**
+     * Converts this {@code Option} to a {@link Try}.
+     *
+     * @return {@code Try.success(get())} if this is defined, otherwise {@code Try.failure(new NoSuchElementException())}
+     */
+    default Try<T> toTry() {
+        return isEmpty() ? Try.failure(new NoSuchElementException()) : Try.success(get());
+    }
+
+    /**
+     * Converts this {@code Option} to a {@link Try}.
+     *
+     * @param ifEmpty an exception supplier
+     * @return {@code Try.success(get())} if this is defined, otherwise {@code Try.failure(ifEmpty.get()}
+     * @throws NullPointerException if the given supplier {@code ifEmpty} is null
+     */
+    default Try<T> toTry(Supplier<? extends Throwable> ifEmpty) {
+        Objects.requireNonNull(ifEmpty, "ifEmpty is null");
+        return isEmpty() ? Try.failure(ifEmpty.get()) : Try.success(get());
     }
 
     @Override
