+++ vavr/src/test/java/io/vavr/control/EitherLeftProjectionTest.java
@@ -146,18 +146,6 @@ public class EitherLeftProjectionTest extends AbstractValueTest {
         Either.right("1").left().getOrElseThrow(s -> new RuntimeException(s));
     }
 
-    // toOption
-
-    @Test
-    public void shouldConvertLeftProjectionOfLeftToSome() {
-        assertThat(Either.left(1).left().toOption()).isEqualTo(Option.of(1));
-    }
-
-    @Test
-    public void shouldConvertLeftProjectionOfRightToNone() {
-        assertThat(Either.right("x").left().toOption()).isEqualTo(Option.none());
-    }
-
     // toEither
 
     @Test
@@ -188,7 +176,7 @@ public class EitherLeftProjectionTest extends AbstractValueTest {
 
     @Test
     public void shouldFilterSomeOnLeftProjectionOfLeftIfPredicateMatches() {
-        final boolean actual = Either.left(1).left().filter(i -> true).toOption().isDefined();
+        final boolean actual = Either.left(1).left().filter(i -> true).isDefined();
         assertThat(actual).isTrue();
     }
